+++ packages/admin-ui/src/lib/core/src/data/definitions/settings-definitions.ts
@@ -547,6 +547,7 @@ export const GET_SERVER_CONFIG = gql`
                     name
                     to
                 }
+                permittedAssetTypes
                 customFieldConfig {
                     Address {
                         ...CustomFields