+++ packages/asset-server-plugin/e2e/graphql/generated-e2e-asset-server-plugin-types.ts
@@ -3235,6 +3235,7 @@ export type SearchResultSortParameter = {
 export type ServerConfig = {
     __typename?: 'ServerConfig';
     orderProcess: Array<OrderProcessState>;
+    permittedAssetTypes: Array<Scalars['String']>;
     customFieldConfig: CustomFields;
 };